style(conf): reconfigure `eslint` for entire proj
Add JSON extension files and `.remarkrc` to list of files eslint parses.

Configure eslint to use the typescript parser for only typescript files,
while also applying the proper rules and formatting to only the files
that are necessary to be evaluated.

Also adjust the ignore list patterns to include hidden files and ignore
special (package-lock) files