test(sys-results): added a `safety.py` potential report
Potential next test case to handle a different scanner output as
compared to npm.  The problem is there is not a module that actually
builds a gitlab schema yet from the `safety.py --json` output yet. But
ideally it would look something like this. Might need to submit issue/
enhancement request to safety.py in order to get them to provide
additional data fields in the output of their json format.